+++ library/src/test/java/com/getstream/sdk/chat/GetDeletedOrMentionedTextTest.java
@@ -86,75 +86,4 @@ public class GetDeletedOrMentionedTextTest {
         text = "@@@Adrian a @@This is @A";
         assertEquals("@@@Adrian a @@This is @Adrian", StringUtility.convertMentionedText(text, userName));
     }
-
-    @Test
-    void containMarkdownTextText() {
-        String text = "TestMarkDown";
-        assertFalse(StringUtility.containsMarkdown(text));
-        text = "*TestMarkDown";
-        assertFalse(StringUtility.containsMarkdown(text));
-        text = "**TestMarkDown";
-        assertFalse(StringUtility.containsMarkdown(text));
-        text = "#TestMarkDown";
-        assertFalse(StringUtility.containsMarkdown(text));
-        text = "##TestMarkDown";
-        assertFalse(StringUtility.containsMarkdown(text));
-        // Emphasis
-        text = "*TestMarkDown*";
-        assertTrue(StringUtility.containsMarkdown(text));
-        text = "_TestMarkDown_";
-        assertTrue(StringUtility.containsMarkdown(text));
-        // Strong emphasis
-        assertTrue(StringUtility.containsMarkdown(text));
-        text = "**TestMarkDown**";
-        assertTrue(StringUtility.containsMarkdown(text));
-        text = "__TestMarkDown__";
-        assertTrue(StringUtility.containsMarkdown(text));
-        // Strike-through
-        text = "~~TestMarkDown~~";
-        assertTrue(StringUtility.containsMarkdown(text));
-        // Headers
-        text = "# TestMarkDown";
-        assertTrue(StringUtility.containsMarkdown(text));
-        text = "## TestMarkDown";
-        assertTrue(StringUtility.containsMarkdown(text));
-        text = "### TestMarkDown";
-        assertTrue(StringUtility.containsMarkdown(text));
-        // Links
-        text = "[GetStream](https://getstream.io/)";
-        assertTrue(StringUtility.containsMarkdown(text));
-        // Images
-        text = "![Minion](https://octodex.github.com/images/minion.png)";
-        assertTrue(StringUtility.containsMarkdown(text));
-        // Thematic break
-        text = "___  ";
-        assertTrue(StringUtility.containsMarkdown(text));
-        text = "---  ";
-        assertTrue(StringUtility.containsMarkdown(text));
-        text = "***  ";
-        assertTrue(StringUtility.containsMarkdown(text));
-        // Code
-        text = "`code`";
-        assertTrue(StringUtility.containsMarkdown(text));
-        text = "// Some comments\n" +
-                "line 1 of code\n" +
-                "line 2 of code\n" +
-                "line 3 of code";
-        assertTrue(StringUtility.containsMarkdown(text));
-
-        text = "```\n" +
-                "Sample text here...\n" +
-                "```";
-        assertTrue(StringUtility.containsMarkdown(text));
-        text = "``` js\n" +
-                "var foo = function (bar) {\n" +
-                "  return bar++;\n" +
-                "};\n" +
-                "\n" +
-                "console.log(foo(5));\n" +
-                "```";
-        assertTrue(StringUtility.containsMarkdown(text));
-        // Blockquotes
-
-    }
 }
